Make Contact polymorphic
This removes the entire concept of `Heya::Contact`, leaving contact
storage up to the user.

I think this opens up some interesting possibilities; for instance, you
could have campaigns for different models.

Not controlling what a "contact" really is (beyond a duck type) should
also keep the system more flexible. By default it uses the `User` class,
which should "just work" in most Rails apps. If you want to be able to
communicate with users *before* they sign up, you could create your own
`Contact` class with an optional relation to users. (This is what I'd
like to do in the Honeybadger app eventually.) We could also provide a
separate engine to do that, if it's a common enough use case.