Ordered sequences (#9)
* Execute sequences in order

Here are two approaches to campaigns:

1. "Smart Campaigns" (Intercom): each message is sent when the
   conditions match. Messages may be sent out of order as a result.
   This style is good for campaigns where each email is dependent on the
   user completing certain actions. Its main drawback is that it's more
   complicated; it requires additional rules for when someone should
   exit a campaign (since the total duration of the campaign is
   unknown). It's also not really a "sequence" as it does not
   necessarily execute the steps in order.

2. Sequence Campaigns (Drip, ConvertKit): each message is sent in order.
   If excluded by conditions, the message is skipped. Once a message is
   skipped it will never be sent, even if conditions match in the
   future (or possibly if the messages are reordered).

Heya is already designed to send messages in steps, with a custom
duration between each step. I think the sequence approach (#2) makes
more sense for us, and it maps to my brain better than Intercom's.

This commit adds a query condition to skip messages when the conditions
don't match. This approach to scheduling is novel and may or may not
work long term. There may be edge cases. :)

* Contacts exit campaign after last message

* Tweak query constant names