Add a base segment to campaigns (#10)
This is a nice way to drill down on who should receive the emails in
your campaign. For example:

```ruby
class SignupFirstRunCampaign < Heya::Campaigns::Base
  default wait: 1.days

  segment { signups }

  step :welcome, wait: 0.days,
    subject: "Welcome to Honeybadger! 👋"

  step :install_1, segment: -> { where.not(id: installed) },
    subject: "How many Honeybadgers does it take?"
end
```

When added to this campaign, contacts who match the "signups" segment
will receive messages. Step-level segments are merged, so that the
"install_1" step will be sent to signups who haven't installed the code
yet. If a contact ever leaves the signups segment while the campaign is
in progress, they will naturally leave the campaign.