Refactor Actions API (#41)
This allows companion gems to define actions like this:

```ruby
module Heya
  module Campaigns
    module Actions
      class Messenger < Action
        def build
          inbox = step.properties.fetch("inbox", :default)
          body = step.properties.fetch("body")
          Heya::Messenger.get_inbox(user, inbox).build(
            from: step.from,
            body: body
          )
        end
      end
    end
  end
end
```

The object built by `Action#build` must respond to a method `#deliver`
(similar to ActionMailer). This allows Heya to build the message object
and give it to the user for unit testing, but deliver it in the
background via ActiveJob (which now happens on a dedicated queue).

The idea is that different actions build different messages, but the
messages are all tested and processed the same.

A side-effect of this change is that block actions are processed in the
background instead of in the scheduler (meaning you don't have to worry
about scheduling your own jobs). I decided that it's simpler to treat
all actions (including blocks) as the same, for now. I may rethink block
actions in the future.