Add optional wait argument to campaign generator
This allows me to set the wait time for each message in the generator
arguments, i.e.:

  rails g heya:campaign Onboarding first:0 second:2.days third

This assumes that anything after the optional : separator is a valid
Ruby duration (which gets inserted as an option in the campaign file).

This is especially useful when you want to send the first message
immediately, as all messages default to a wait time of 2 days.