Inject CTE and refactor to scopes
This is a bit unorthodox, but given ActiveRecord's lack of support for
Common Table Expressions (CTEs, otherwise known as the WITH statement),
it's the only option that lets me use a first-class CTE in my native
ActiveRecord scopes. The primary benefit is that this will make it much
easier for users to query for upcoming messages via the familiar
ActiveRecord interface; as far as I can tell, the alternatives are to
provide an Arel relation, or raw SQL queries--the latter being less
versatile, and the former being more versatile but less familiar.

With this approach, querying for the next message becomes trivial:

Heya::CampaignMembership.upcoming.first

This approach also simplifies Heya's internal code (it will be easier to
add a dashboard in the future, for instance).

Inspired by https://github.com/kmurph73/ctes_in_my_pg (which is inspired
by https://github.com/DavyJonesLocker/postgres_ext).